+++ internal/embed/math.go
@@ -1,0 +1,5 @@
+package embed
+
+import "math"
+
+func _sqrt(x float64) float64 { return math.Sqrt(x) }